Configured Cmake and refactored some code
